+++ app/Http/Controllers/AuthController.php
@@ -2,15 +2,23 @@
 
 namespace App\Http\Controllers;
 
+use App\Kafka\Producer\BaseProducer;
 use App\Models\User;
-use Auth;
+use App\Notifications\VerifyEmail;
 use Illuminate\Foundation\Auth\EmailVerificationRequest;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Hash;
-use Validator;
+use Illuminate\Support\Facades\Validator;
+use Symfony\Component\HttpFoundation\RedirectResponse;
 
 class AuthController extends Controller
 {
+    public function __construct(
+        private readonly BaseProducer $producer
+    ) {
+    }
+
     public function resetPassword(Request $request)
     {
         $user = User::where('email', $request->email)->first();
@@ -39,6 +47,18 @@ class AuthController extends Controller
             return redirect()->route('site.my-courses');
         }
         return view('auth.verify-email', ['user' => $user]);
+    }
+
+    public function sendEmailVerificationNotification(Request $request): RedirectResponse
+    {
+        $user = Auth::user();
+        if ($user->hasVerifiedEmail()) {
+            return redirect()->route('site.my-courses');
+        }
+
+        $user->notify(new VerifyEmail($this->producer));
+
+        return back()->with('resent', true);
     }
 
     public function successVerification()